refactor: move tryFormatting to separate file
